+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 2. [](#improved)
     * Improved `Filesystem::basename()` and `Filesystem::pathinfo()` to be unicode-safe
     * Made path handling unicode-safe, use new `Utils::basename()` and `Utils::pathinfo()` everywhere
+    * Fixed MimeType for `gzip` (`application/x-gzip`)
 
 # v1.7.28
 ## 01/24/2022