+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # v1.7.29
-## mm/dd/2022
+## 01/28/2022
 
 1. [](#new)
     * Added support for registering assets from `HtmlBlock`